iOS 13 fix
Dropcaps appear larger than they should be for certain fonts in iOS 13 Safari.

iOS 13 Safari appears to render text poorly when using the textBaseline property. Using the "top" value means the capital E is rendered out of intended position and makes the resulting dropcap appear larger than it should be.

When using "middle" for textBaseline, it's still slightly out of alignment but close enough for this library to work correctly. Of course, this means having to set the render to begin half way down the canvas element.